refactor teaminvitelist code and fix leaving a team
